painter: [refactor] rename painter.ts to index.ts
