+++ src/components/painter/index.ts
